requirements.txt
pandas
matplotlib
scikit-learn
joblib
pillow
opencv-python
mss
pynput
ultralytics
pytesseract
notebook